Wrap-format-response obeys Accept headers.
Wrap-format-response takes a sequence of encoders and selects the
preferred one according to the Accept header. If no suitable encoder
is found, 406 Not Acceptable is returned. Content-Type of the encoded
string is given with the encoder, so wrap-format-response no longer
takes a type parameter.

Make-encoder takes an encoder function and Content-Type of the encoded
string. Returns a map suitable for wrap-format-response.

Wrap-<type>-response function uses the new wrap-format-response but
the signature is unchanged.

Wrap-restful-response uses the preferred encoder according to Accept,
or returns 406 Not Acceptable. :default should be a map returned by
make-encoder.