Update README.md
Add a Oudin Coil pic: An Oudin coil, a resonant transformer circuit invented by Paul Marie Oudin in 1892, which generates high voltage, low current, high frequency AC electricity, used in the obsolete medical field of electrotherapy around the turn of the 20th century. The high voltage terminal at the top of the coil (B) was connected through a wire to a handheld electrode which produced luminous streamer arcs, which were applied to the body to treat various medical conditions. It was very similar to a Tesla coil, the only difference being that in the Oudin the output voltage was taken across both the primary and secondary windings of the coil in series, like an autotransformer.