Add ability to test things in Cursive IDE
Having the same project coordinates (metosin/muuntaja) for the
parent project and a module clashed in Cursive.

Also, some tests require jsonista namespaces without stating that
dependency, obviously leaving things to implicit dependencies
(from module metosin/muuntaja). This makes it hard to start a REPL
in Cursive with the proper settings.

With this commit, I was able to start a Cursive based REPL where I
was able to run both the tests and (re-)load the `muuntaja.core` ns).
This increases (my) development speed.