Fix reading empty bodies
If there is nothing to decode, it is okay to return nil. That's not
a failure to decode. This problem came up when testing HEAD requests,
which are body-less by default. If we still run it through middleware
to decode the body, that should not fail. It's not super-elegant, as
we already know we do not care for the body anyhow (or at least we
shouldn't care), but it should not fail. Also, it might arise in other
situations, too.