Support ‘application/x-www-form-urlencoded’
This implements encoding/decoding for the
`application/x-www-form-urlencoded` media-type which is the default
format sent by browsers when submitting HTML forms using the POST
method.

The implementation relies on the `ring.util.codec` library to perform
both the encoding and decoding. This is the same library that is used
by the `ring.middleware.params` alternative solution for handling both
query params and body params encoded with the
`application/x-www-form-urlencoded` media-type.